fix: fix type of label to int64 (#1353)

+++ ppcls/data/dataloader/imagenet_dataset.py
@@ -34,5 +34,5 @@ class ImageNetDataset(CommonDataset):
             for l in lines:
                 l = l.strip().split(" ")
                 self.images.append(os.path.join(self._img_root, l[0]))
-                self.labels.append(int(l[1]))
+                self.labels.append(np.int64(l[1]))
                 assert os.path.exists(self.images[-1])
